[#390] refactor: withdrawalType 파일 위치 이동
