+++ src/main/java/com/jeju/nanaland/domain/member/entity/enums/WithdrawalType.java
@@ -1,4 +1,4 @@
-package com.jeju.nanaland.domain.member.entity;
+package com.jeju.nanaland.domain.member.entity.enums;
 
 public enum WithdrawalType {
   INSUFFICIENT_CONTENT, INCONVENIENT_SERVICE, INCONVENIENT_COMMUNITY, RARE_VISITS
